refactored stats system
