+++ Assets/Prefabs/Enemy/Dasher.prefab
@@ -243,7 +243,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a4af50f36cf2a694991d9ecd6288232e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  enemyData: {fileID: 11400000, guid: dc4018f4077fdb149bdc390719c85105, type: 2}
+  data: {fileID: 11400000, guid: dc4018f4077fdb149bdc390719c85105, type: 2}
   target: {fileID: 0}
 --- !u!195 &3781740808952231667
 NavMeshAgent:
