+++ Assets/Scripts/Data/CombatEntityData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 154ea85f5cf696741a18c83ae701304f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
