+++ Assets/Scripts/Entities/CombatEntity.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 944990fe24cf7214d968aea7b07dc87e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
