+++ Assets/Scripts/Entities/Enemies/Dasher.cs
@@ -4,20 +4,16 @@ using UnityEngine;
 
 public class Dasher : Enemy
 {
-    #region States
-    
-    #endregion
-
     protected override void Awake()
     {
         base.Awake();
-        IdleState = new DasherIdleState(this, StateMachine, enemyData, "Idle");
-        PatrolState = new DasherPatrolState(this, StateMachine, enemyData, "Move");
-        AlertedState = new DasherAlertedState(this, StateMachine, enemyData, "Alerted");
-        AgroState = new DasherAgroState(this, StateMachine, enemyData, "Move");
-        AttackState = new DasherAttackState(this, StateMachine, enemyData, "Attack");
-        CCState = new DasherCCState(this, StateMachine, enemyData, "Idle");
-        DeathState = new DasherDeathState(this, StateMachine, enemyData, "Death");
+        IdleState = new DasherIdleState(this, StateMachine, (EnemyData)data, "Idle");
+        PatrolState = new DasherPatrolState(this, StateMachine, (EnemyData)data, "Move");
+        AlertedState = new DasherAlertedState(this, StateMachine, (EnemyData)data, "Alerted");
+        AgroState = new DasherAgroState(this, StateMachine, (EnemyData)data, "Move");
+        AttackState = new DasherAttackState(this, StateMachine, (EnemyData)data, "Attack");
+        CCState = new DasherCCState(this, StateMachine, (EnemyData)data, "Idle");
+        DeathState = new DasherDeathState(this, StateMachine, (EnemyData)data, "Death");
     }
 
     protected override void Start()
